refactor searchworks subjects methods into a separate file
